// src/app/models/DTOs/GlassesSatisDetailDto.ts
export default interface GlassesSatisDetaliDto {
    
        glassesId: number
        customerId: number
        adi: string
        soyadi: string
        kimlikNo: string
        yKimlikN:string
        glassBrand: string
        glassCategory: string
        glassType:string
        rsph:string
        rcyl:string
        raks:string
        lsph:string
        lcyl:string
        laks:string
        frameBrand:string 
        frameModel: string
        frameEcartman:string
        frameColor: string
        frameSerialNumber:string 
        prescriptNo: string
        prescriptDate: Date
        isEntered:string
        gozlukNot: string
        satisTarihi:string
        tutar:number
}